Added bgdat.py, just read support
This commit just adds bgdat.py, with a single function called `read_bgdat(file)`.
Basically, what it does is return a list of dictionaries (every dictionary is about a tile added in Reggie). It returns 3 values: the list of dictionaries, the full file readed, and the number of extra bits (FF).